Extend the form field renderer
* a new type `error` has been added. The reason for the new type is that we can't access to the
error values of a field out of the renderer component (https://github.com/erikras/redux-form/issues/396).
We used the fact that fields with same name have the all attributes in the sync. We are using one
field to render the input form and the other with the same name to render the error.
The width of a numberical field can be too small to show the validation message. In order to solve the
issue, the matching error field has been added. The attribute suppressErrors will disable the rendering of
the error message (the error will be shown by the error field). This is useful only if we, for example,
need to collect all errors on the top of the screen or if we have an issue with the width of the input filed:

+--------------------------------------------------------------------+
| +-------------+--------------------------------------------------+ |
| | input:num   | label                                            | |
| | ----------- |                                                  | |
| | input:error |  <-- suppressed rendering                        | |
| +-------------+--------------------------------------------------+ |
+--------------------------------------------------------------------+
| +----------------------------------------------------------------+ |
| | input:error                                                    | |
| +----------------------------------------------------------------+ |
+--------------------------------------------------------------------+